Trait decorum::cmp::FloatOrd

pub trait FloatOrd {
    fn float_cmp(&self, other: &Self) -> Ordering;
}

Total ordering of primitive floating-point types.

FloatOrd expresses the total ordering:

$$-\infin<\cdots<0<\cdots<\infin<\text{NaN}$$

This trait can be used to compare primitive floating-point types without the need to wrap them within a proxy type. See the module documentation for more about the ordering used by FloatOrd and proxy types.

Required Methods

fn float_cmp(&self, other: &Self) -> Ordering