decorum

Trait Float

pub trait Float:
    Encoding
    + Infinite
    + IntrinsicOrd
    + Nan
    + Real { }

Floating-point representations.

Types that implement this trait are represented using IEEE-754 encoding and expose the details of that encoding, including infinities, NaN, and operations on real numbers. This trait is implemented by primitive floating-point types and the Total proxy type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.