Struct decorum::ConstrainedFloat

#[repr(transparent)]
pub struct ConstrainedFloat<T, P> { /* private fields */ }

Floating-point proxy that provides a total ordering, equivalence, hashing, and constraints.

ConstrainedFloat wraps primitive floating-point types and provides implementations for numeric traits using a total ordering, including Ord, Eq, and Hash. ConstrainedFloat supports various constraints on the class of values that may be represented and panics if these constraints are violated.

This type is re-exported but should not (and cannot) be used directly. Use the type aliases Total, NotNan, and Finite instead.

Total Ordering

All proxy types use the following total ordering:

$$-\infin<\cdots<0<\cdots<\infin<\text{NaN}$$

See the cmp module for a description of the total ordering used to implement Ord and Eq.

Constraints

Constraints restrict the set of values that a proxy may take by disallowing certain classes or subsets of those values. If a constraint is violated (because a proxy type would need to take a value it disallows), the operation panics.

Constraints may disallow two broad classes of floating-point values: infinities and NaNs. Constraints are exposed by the Total, NotNan, and Finite type definitions. Note that Total uses a unit constraint, which enforces no constraints at all and never panics.

Implementations

impl<T, P> ConstrainedFloat<T, P> where
T: Float + Primitive,
P: Constraint<T>,

pub fn from_inner(value: T) -> Self

Converts a primitive floating-point value into a proxy.

The same behavior is provided by an implementation of the From trait.

Panics

This conversion and the implementation of the From trait will panic if the primitive floating-point value violates the constraints of the proxy.

Examples

Converting primitive floating-point values into proxies:

use decorum::R64;

fn f(x: R64) -> R64 {
    x * 2.0
}

// Conversion using `from_inner`.
let y = f(R64::from_inner(2.0));
// Conversion using `From`/`Into`.
let z = f(2.0.into());

Performing a conversion that panics:

use decorum::R64;

// `R64` does not allow `NaN`s, but `0.0 / 0.0` produces a `NaN`.
let x = R64::from_inner(0.0 / 0.0); // Panics.

pub fn into_inner(self) -> T

Converts a proxy into a primitive floating-point value.

Examples

Converting a proxy into a primitive floating-point value:

use decorum::R64;

fn f() -> R64 {
    // ...
}

let x: f64 = f().into_inner();

pub fn from_subset<Q>(other: ConstrainedFloat<T, Q>) -> Self where
Q: Constraint<T> + SubsetOf<P>,

Converts a proxy into another proxy that is capable of representing a superset of the values that are members of its constraint.

Examples

Converting between compatible proxy types:

use decorum::{N64, R64};
use num::Zero;

let x = R64::zero();
let y = N64::from_subset(x);

pub fn into_superset<Q>(self) -> ConstrainedFloat<T, Q> where
Q: Constraint<T> + SupersetOf<P>,

Converts a proxy into another proxy that is capable of representing a superset of the values that are members of its constraint.

Examples

Converting between compatible proxy types:

use decorum::{N64, R64};
use num::Zero;

let x = R64::zero();
let y: N64 = x.into_superset();